requirements.txt
Flask
line-bot-sdk
python-dotenv
gunicorn
requests